!!! Everything is broken but it's fine
I just refactored the database and successfully, and saved a backup to drive.

I've updated the data models for the (hopefully) final database design for v1.  Now it's all cleaning up the C# layers.  Anything that doesn't belong will be pruned, hidden classes will need to be deleted.  We'll see how quickly that goes.